+++ frontend/src/pages/Tenant.tsx
@@ -3,22 +3,34 @@ import {
     Center,
     Checkbox,
     Flex,
+    Grid,
+    GridItem,
+    Select,
     Spinner,
     Stack,
     Text,
+    useDisclosure,
 } from '@chakra-ui/react';
 import { IconArrowBack } from '@tabler/icons';
-import { useCallback, useMemo } from 'react';
+import { isEmpty } from 'ramda';
+import { useCallback, useMemo, useRef, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import useSWR from 'swr';
+import ConfirmModal from '../components/Modals/ConfirmModal';
 import Breadcrumbs from '../components/Navigation/Breadcrumbs';
 import PageContainer from '../components/PageContainer';
+import PropertyCard from '../components/Property/PropertyCard';
 import { useNotification } from '../hooks/useNotification';
+import {
+    PropertyApiService,
+    propertyApiService,
+} from '../services/api/PropertyApiService';
 import { rentApiService, RentApiService } from '../services/api/RentApiService';
 import {
     tenantApiService,
     TenantApiService,
 } from '../services/api/TenantApiService';
+import { Property, PropertyList } from '../types/property';
 import { Rent, RentList } from '../types/rent';
 import { Tenant } from '../types/tenant';
 
@@ -31,6 +43,7 @@ function TenantPage() {
         data: tenant,
         error: tenantError,
         isValidating: isValidatingTenant,
+        mutate: mutateTenant,
     } = useSWR<Tenant>(
         TenantApiService.getTenantPath(Number(id)),
         tenantApiService.getTenant,
@@ -48,6 +61,30 @@ function TenantPage() {
     );
 
     const {
+        data: allPropertiesData,
+        error: allPropertiesError,
+        isValidating: isValidatingAllProperties,
+    } = useSWR<PropertyList>(
+        PropertyApiService.listPropertiesPath,
+        propertyApiService.getProperties,
+    );
+
+    const allProperties = useMemo(
+        () => allPropertiesData?.properties,
+        [allPropertiesData],
+    );
+    const isAllPropertiesError = useMemo(
+        () => allPropertiesError !== undefined,
+        [allPropertiesError],
+    );
+    const isLoadingAllProperties = useMemo(
+        () =>
+            allProperties !== undefined ||
+            (isValidatingAllProperties && isAllPropertiesError),
+        [allProperties, isValidatingAllProperties, isAllPropertiesError],
+    );
+
+    const {
         data: rentData,
         error: rentError,
         isValidating: isValidatingRent,
@@ -57,6 +94,17 @@ function TenantPage() {
         rentApiService.getRents,
     );
 
+    const {
+        isOpen: isConfirmAddModalOpen,
+        onOpen: openConfirmAddModal,
+        onClose: closeConfirmAddModal,
+    } = useDisclosure();
+    const {
+        isOpen: isConfirmRemoveModalOpen,
+        onOpen: openConfirmRemoveModal,
+        onClose: closeConfirmRemoveModal,
+    } = useDisclosure();
+
     const isLoadingRents = useMemo(
         () =>
             rentData === undefined ||
@@ -65,6 +113,13 @@ function TenantPage() {
     );
     const isRentError = useMemo(() => rentError !== undefined, [rentError]);
     const rents = useMemo(() => rentData?.rents, [rentData]);
+    const properties = useMemo(() => tenant?.properties, [tenant]);
+
+    const [selectedProperty, setSelectedProperty] = useState<Property | null>(
+        null,
+    );
+
+    const selectRef = useRef<HTMLSelectElement>(null);
 
     const breadcrumbs = useMemo(
         () => [
@@ -99,8 +154,82 @@ function TenantPage() {
         }
     }, []);
 
+    const onTenantAddSubmit = useCallback(async () => {
+        if (tenant && selectedProperty && selectRef.current) {
+            try {
+                await propertyApiService.addTenantToProperty(
+                    PropertyApiService.addTenantToPropertyPath(
+                        selectedProperty.id,
+                        tenant.id,
+                    ),
+                );
+
+                showSuccess(
+                    'Tenant added',
+                    `${tenant.name} was added to ${selectedProperty.name}.`,
+                );
+            } catch (e) {
+                showError(
+                    'Error',
+                    'An error occured while trying to add the tenant to the property.',
+                );
+            } finally {
+                setSelectedProperty(null);
+                mutateTenant();
+                selectRef.current.value = '';
+                closeConfirmAddModal();
+            }
+        }
+    }, [tenant, selectedProperty]);
+
+    const onTenantRemoveSubmit = useCallback(async () => {
+        if (selectedProperty && tenant) {
+            try {
+                await propertyApiService.removeTenantFromProperty(
+                    PropertyApiService.removeTenantFromPropertyPath(
+                        selectedProperty.id,
+                        tenant.id,
+                    ),
+                );
+
+                showSuccess(
+                    'Tenant removed',
+                    `${tenant.name} was removed from ${selectedProperty.name}`,
+                );
+            } catch (e) {
+                showError(
+                    'Error',
+                    `An error occurred while trying to remove ${tenant.name} from ${selectedProperty.name}`,
+                );
+            } finally {
+                setSelectedProperty(null);
+                closeConfirmRemoveModal();
+                mutateTenant();
+            }
+        }
+    }, [tenant, selectedProperty]);
+
+    const onTenantAddCancel = useCallback(() => {
+        setSelectedProperty(null);
+        closeConfirmAddModal();
+    }, []);
+
     return (
         <Box>
+            <ConfirmModal
+                isOpen={isConfirmAddModalOpen}
+                title={`Add ${tenant?.name} to ${selectedProperty?.name}?`}
+                message={`Are you sure you want to add ${tenant?.name} to ${selectedProperty?.name}?`}
+                onConfirm={onTenantAddSubmit}
+                onCancel={onTenantAddCancel}
+            />
+            <ConfirmModal
+                isOpen={isConfirmRemoveModalOpen}
+                title={`Remove ${tenant?.name} from ${selectedProperty?.name}?`}
+                message={`Are you sure you want to remove ${tenant?.name} from ${selectedProperty?.name}?`}
+                onConfirm={onTenantRemoveSubmit}
+                onCancel={closeConfirmRemoveModal}
+            />
             <Breadcrumbs items={breadcrumbs} />
             <PageContainer>
                 {isLoadingTenant && (
@@ -128,31 +257,90 @@ function TenantPage() {
                         {!isLoadingRents && !isRentError && rents && (
                             <Stack spacing={1}>
                                 <Text fontSize="xl">Rents:</Text>
-                                {rents.length === 0 && (
-                                    <Text>No rents found</Text>
-                                )}
-                                {rents.length > 0 &&
-                                    rents.map((rent) => {
-                                        return (
-                                            <Flex
-                                                key={rent.due_date}
-                                                gap={2}
-                                                alignItems="center">
-                                                <Text>
-                                                    {new Date(
-                                                        rent.due_date,
-                                                    ).toDateString()}
-                                                </Text>
-                                                <Checkbox
-                                                    checked={rent.paid}
-                                                    onChange={() =>
-                                                        onRentStatusUpdate(rent)
-                                                    }
-                                                />
-                                            </Flex>
-                                        );
-                                    })}
+                                {isEmpty(rents) && <Text>No rents found</Text>}
+                                {rents.map((rent) => {
+                                    return (
+                                        <Flex
+                                            key={rent.due_date}
+                                            gap={2}
+                                            alignItems="center">
+                                            <Text>
+                                                {new Date(
+                                                    rent.due_date,
+                                                ).toDateString()}
+                                            </Text>
+                                            <Checkbox
+                                                checked={rent.paid}
+                                                onChange={() =>
+                                                    onRentStatusUpdate(rent)
+                                                }
+                                            />
+                                        </Flex>
+                                    );
+                                })}
                             </Stack>
+                        )}
+                        {isLoadingTenant && (
+                            <Center>
+                                <Spinner size="lg" />
+                            </Center>
+                        )}
+                        {!isLoadingTenant && !isTenantError && properties && (
+                            <Stack spacing={1}>
+                                <Text fontSize="xl">Properties:</Text>
+                                {isEmpty(properties) && (
+                                    <Text>No properties found</Text>
+                                )}
+                                <Grid templateColumns="repeat(3, 1fr)" gap={6}>
+                                    {properties.map((property) => (
+                                        <GridItem key={property.id}>
+                                            <PropertyCard
+                                                property={property}
+                                                onDeleteClick={() => {
+                                                    setSelectedProperty(
+                                                        property,
+                                                    );
+                                                    openConfirmRemoveModal();
+                                                }}
+                                            />
+                                        </GridItem>
+                                    ))}
+                                </Grid>
+                            </Stack>
+                        )}
+                        {allProperties && properties && (
+                            <Flex gap={2} direction="column">
+                                <Text fontSize="xl">Add property:</Text>
+                                <Select
+                                    placeholder="Properties..."
+                                    ref={selectRef}
+                                    onChange={(e) => {
+                                        setSelectedProperty(
+                                            allProperties.find(
+                                                (property) =>
+                                                    property.id ===
+                                                    Number(e.target.value),
+                                            ) ?? null,
+                                        );
+
+                                        openConfirmAddModal();
+                                    }}
+                                    isDisabled={
+                                        isEmpty(allProperties) ||
+                                        selectedProperty !== null
+                                    }>
+                                    {allProperties.map((property) => (
+                                        <option
+                                            key={property.id}
+                                            value={property.id}
+                                            disabled={properties
+                                                .map((p) => p.id)
+                                                .includes(property.id)}>
+                                            {property.name}
+                                        </option>
+                                    ))}
+                                </Select>
+                            </Flex>
                         )}
                     </Stack>
                 )}