+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Marking rent status via UI ([@kovaj024](https://github.com/kovaj024))
 - Settings page to set preferences ([@zvonimirr](https://github.com/zvonimirr))
 - Remove React Select Currency ([@zvonimirr](https://github.com/zvonimirr))
+- Allow adding/removing tenant to a property from the tenant overview page ([@kovaj024](https://github.com/kovaj024))
 
 ## 0.1.0